requirements.txt
# requirements.txt
flask
openai
python-dotenv
